pyproject.toml
[build-system]
requires = ["hatchling>=1.21.1"]
build-backend = "hatchling.build"

[project]
name = "songviz"
version = "0.0.0"
description = "Generate a music-reactive video from an audio file."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [{name = "SongViz Contributors"}]
dependencies = [
  "numpy>=1.24",
  "librosa>=0.10.1",
  "soundfile>=0.12",
  "Pillow>=10.0",
  "imageio-ffmpeg>=0.4.9",
]

[project.optional-dependencies]
test = [
  "pytest>=7.4",
]
stems = [
  "demucs",
  "torchcodec",
]
notes = [
  "basic-pitch",
  # basic-pitch currently pulls native wheels (e.g., tflite-runtime) that may not
  # be compatible with NumPy 2.x on all platforms.
  "numpy<2",
]
lyrics = [
  "openai-whisper",
  "mutagen>=1.47",
  "stable-ts",
  "pyyaml>=6.0",
]
lyricsx = [
  "whisperx",
]
viz = [
  "matplotlib",
]

[project.scripts]
songviz = "songviz.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["songviz"]

[tool.pytest.ini_options]
testpaths = ["tests"]
cache_dir = ".songviz/pytest_cache"
